Refactor List Command Class
