+++ src/main/java/seedu/address/logic/commands/client/ListClientCommand.java
@@ -1,16 +1,18 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.client;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_CLIENTS;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 /**
  * Lists all Clients in the address book to the user.
  */
-public class ListCommand extends Command {
+public class ListClientCommand extends Command {
 
-    public static final String COMMAND_WORD = "list";
+    public static final String COMMAND_WORD = "clist";
 
     public static final String MESSAGE_SUCCESS = "Listed all Clients";
